Anonymous subclass for install_name_tool step
