--- src/com/facebook/buck/apple/InstallNameToolStep.java
@@ -1,75 +1,0 @@
-/*
- * Copyright 2015-present Facebook, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License. You may obtain
- * a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- */
-
-package com.facebook.buck.apple;
-
-import com.facebook.buck.io.ProjectFilesystem;
-import com.facebook.buck.shell.ShellStep;
-import com.facebook.buck.step.ExecutionContext;
-import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
-
-import java.nio.file.Path;
-import java.util.List;
-
-/**
- * {@link ShellStep} implementation which invokes Apple's {@code install_name_tool}
- * utility to set @rpaths on compiled binaries.
- */
-class InstallNameToolStep extends ShellStep {
-
-  private final ProjectFilesystem filesystem;
-  private final ImmutableMap<String, String> environment;
-  private final ImmutableList<String> installNameToolCommand;
-  private final ImmutableList<String> additionalParams;
-  private final Path binary;
-
-  public InstallNameToolStep(
-      ProjectFilesystem filesystem,
-      ImmutableMap<String, String> environment,
-      List<String> installNameToolCommand,
-      List<String> additionalParams,
-      Path input) {
-    super(filesystem.getRootPath());
-    this.filesystem = filesystem;
-    this.environment = environment;
-    this.installNameToolCommand = ImmutableList.copyOf(installNameToolCommand);
-    this.additionalParams = ImmutableList.copyOf(additionalParams);
-    this.binary = input;
-  }
-
-  @Override
-  protected ImmutableList<String> getShellCommandInternal(ExecutionContext context) {
-    ImmutableList.Builder<String> commandBuilder = ImmutableList.builder();
-
-    commandBuilder.addAll(installNameToolCommand);
-    commandBuilder.addAll(additionalParams);
-    commandBuilder.add(
-        filesystem.resolve(binary).toString());
-
-    return commandBuilder.build();
-  }
-
-  @Override
-  public ImmutableMap<String, String> getEnvironmentVariables(ExecutionContext context) {
-    return environment;
-  }
-
-  @Override
-  public String getShortName() {
-    return "install_name_tool";
-  }
-}
